requirements.txt
discord.py>=2.4.0
ezjsonpy==1.0.5
loguru>=0.7.3
python-dotenv>=1.0.1
requests>=2.32.3